+++ lib/seo_cache/page_render.rb
@@ -12,16 +12,8 @@ module SeoCache
       return @driver.page_source
     rescue StandardError => error
       SeoCache.log_error(error.message)
-    ensure
-      @driver&.quit
-    end
-
-    def persistent_get(url)
-      @driver.get(url)
-
-      return @driver.page_source
-    rescue StandardError => error
-      SeoCache.log_error(error.message)
+    # ensure
+    #   @driver&.quit
     end
 
     def close_connection
@@ -37,7 +29,6 @@ module SeoCache
 
       Selenium::WebDriver::Chrome.path = SeoCache.chrome_path if SeoCache.chrome_path
 
-      client          = ::Selenium::WebDriver::Remote::Http::Persistent.new
       browser_options = ::Selenium::WebDriver::Chrome::Options.new
       browser_options.args << '--headless'
       browser_options.args << '--disable-gpu'
@@ -45,7 +36,7 @@ module SeoCache
       browser_options.args << '--disable-web-security'
       browser_options.args << '--window-size=1920x1080'
       # browser_options.args << '--remote-debugging-port=3020'
-      @driver = ::Selenium::WebDriver.for(:chrome, options: browser_options, http_client: client)
+      @driver = ::Selenium::WebDriver.for(:chrome, options: browser_options)
     end
   end
 end